+++ src/main/java/mygame/editor/component/SelectComponent.java
@@ -25,7 +25,7 @@ public class SelectComponent extends Component {
     public void draw(GraphicsContext g) {
         if(owner.active) {
             g.setStroke(Color.CYAN);
-            g.strokeRect(-5, -owner.getHeight() - 5, owner.getWidth()+ 10, owner.getHeight()+ 10);
+            g.strokeRect(-5, -owner.getHeight().doubleValue() - 5, owner.getWidth().doubleValue()+ 10, owner.getHeight().doubleValue()+ 10);
             g.stroke();
         }
     }
